blktrace: Add ENOMEM handling in `trace_add_open_close_event()` and its callers
When this function hits ENOMEM, there is no way to notify the callers
that we fail to allocate the memory. This changes the return value of
`trace_add_open_close_event()` to `int` so we can tell the caller if
it hits ENOMEM. Also, fix the callers to handle this error case.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>